+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/rentalreservation/application/listener/DepositHeldEventListener.java
@@ -1,7 +1,7 @@
 package com.nextdoor.nextdoor.domain.rentalreservation.application.listener;
 
 import com.nextdoor.nextdoor.domain.fintech.event.DepositHeldEvent;
-import com.nextdoor.nextdoor.domain.rentalreservation.application.service.RentalService;
+import com.nextdoor.nextdoor.domain.rentalreservation.application.service.RentalSettlementService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
@@ -12,11 +12,11 @@ import org.springframework.transaction.event.TransactionalEventListener;
 @RequiredArgsConstructor
 public class DepositHeldEventListener {
 
-    private final RentalService rentalService;
+    private final RentalSettlementService rentalSettlementService;
 
     @Async("asyncExecutor")
     @TransactionalEventListener(phase = TransactionPhase.BEFORE_COMMIT)
     public void handleDepositHeldEvent(DepositHeldEvent event) {
-        rentalService.updateRentalDepositId(event.getRentalId(), event.getDepositId());
+        rentalSettlementService.updateRentalDepositId(event.getRentalId(), event.getDepositId());
     }
 }
