+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/rentalreservation/application/listener/RemittanceCompletedEventListener.java
@@ -1,7 +1,7 @@
 package com.nextdoor.nextdoor.domain.rentalreservation.application.listener;
 
 import com.nextdoor.nextdoor.domain.fintech.event.RemittanceCompletedEvent;
-import com.nextdoor.nextdoor.domain.rentalreservation.application.service.RentalService;
+import com.nextdoor.nextdoor.domain.rentalreservation.application.service.RentalSettlementService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.scheduling.annotation.Async;
@@ -13,12 +13,12 @@ import org.springframework.transaction.event.TransactionalEventListener;
 @RequiredArgsConstructor
 public class RemittanceCompletedEventListener {
 
-    private final RentalService rentalService;
+    private final RentalSettlementService rentalSettlementService;
     private final SimpMessagingTemplate messagingTemplate;
 
     @Async("asyncExecutor")
     @TransactionalEventListener(phase = TransactionPhase.BEFORE_COMMIT)
     public void handleRemittanceCompletedEvent(RemittanceCompletedEvent remittanceCompletedEvent){
-        rentalService.completeRemittanceProcessing(remittanceCompletedEvent);
+        rentalSettlementService.completeRemittanceProcessing(remittanceCompletedEvent);
     }
 }
